+++ Postgrest/src/commonMain/kotlin/io/github/jan/supabase/postgrest/request/InsertRequest.kt
@@ -1,8 +1,7 @@
-package io.github.jan.supabase.postgrest.request.impl
+package io.github.jan.supabase.postgrest.request
 
 import io.github.jan.supabase.postgrest.query.Count
 import io.github.jan.supabase.postgrest.query.Returning
-import io.github.jan.supabase.postgrest.request.PostgrestRequest
 import io.ktor.http.HttpMethod
 import kotlinx.serialization.json.JsonArray
