+++ Postgrest/src/commonMain/kotlin/io/github/jan/supabase/postgrest/request/SelectRequest.kt
@@ -1,4 +1,4 @@
-package io.github.jan.supabase.postgrest.request.impl
+package io.github.jan.supabase.postgrest.request
 
 import io.github.jan.supabase.postgrest.query.Count
 import io.github.jan.supabase.postgrest.request.PostgrestRequest
