+++ Postgrest/src/commonTest/kotlin/request/InsertRequestTest.kt
@@ -1,9 +1,9 @@
-package request.impl
+package request
 
 import io.github.jan.supabase.postgrest.query.Count
 import io.github.jan.supabase.postgrest.query.Returning
+import io.github.jan.supabase.postgrest.request.InsertRequest
 import io.github.jan.supabase.postgrest.request.PostgrestRequest
-import io.github.jan.supabase.postgrest.request.impl.InsertRequest
 import kotlinx.serialization.json.JsonArray
 import kotlin.test.Test
 import kotlin.test.assertEquals
