+++ Postgrest/src/commonTest/kotlin/request/SelectRequestTest.kt
@@ -1,8 +1,8 @@
-package request.impl
+package request
 
 import io.github.jan.supabase.postgrest.query.Count
 import io.github.jan.supabase.postgrest.request.PostgrestRequest
-import io.github.jan.supabase.postgrest.request.impl.SelectRequest
+import io.github.jan.supabase.postgrest.request.SelectRequest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
